Cut + paste from readwrite to tasks manager file
Cut and paste methods related to crud tasks or running (vs. triggering) a task
from ReadWrite file to TasksManager file
