+++ controller/tasks_manager_test.go
@@ -2,6 +2,7 @@ package controller
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"testing"
 	"time"
@@ -13,6 +14,7 @@ import (
 	mocksS "github.com/hashicorp/consul-terraform-sync/mocks/store"
 	mocksTmpl "github.com/hashicorp/consul-terraform-sync/mocks/templates"
 	"github.com/hashicorp/consul-terraform-sync/state"
+	"github.com/hashicorp/consul-terraform-sync/state/event"
 	"github.com/hashicorp/consul-terraform-sync/templates"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -602,6 +604,438 @@ func TestServer_addTask(t *testing.T) {
 		case <-time.After(time.Second * 5):
 			t.Fatal("did not receive from deleteCh as expected")
 		}
+	})
+}
+
+func TestReadWrite_CheckApply(t *testing.T) {
+	t.Parallel()
+
+	cases := []struct {
+		name          string
+		expectError   bool
+		enabledTask   bool
+		applyTaskErr  error
+		renderTmplErr error
+		taskName      string
+		addToStore    bool
+	}{
+		{
+			"error on driver.RenderTemplate()",
+			true,
+			true,
+			nil,
+			errors.New("error on driver.RenderTemplate()"),
+			"task_render_tmpl",
+			true,
+		},
+		{
+			"error on driver.ApplyTask()",
+			true,
+			true,
+			errors.New("error on driver.ApplyTask()"),
+			nil,
+			"task_apply",
+			true,
+		},
+		{
+			"error creating new event",
+			true,
+			true,
+			nil,
+			nil,
+			"",
+			false,
+		},
+		{
+			"disabled task",
+			false,
+			false,
+			nil,
+			nil,
+			"disabled_task",
+			false,
+		},
+		{
+			"happy path",
+			false,
+			true,
+			nil,
+			nil,
+			"task_apply",
+			true,
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			d := new(mocksD.Driver)
+			drivers := driver.NewDrivers()
+			var task *driver.Task
+			if tc.enabledTask {
+				task = enabledTestTask(t, tc.taskName)
+				d.On("RenderTemplate", mock.Anything).
+					Return(true, tc.renderTmplErr)
+				d.On("ApplyTask", mock.Anything).Return(tc.applyTaskErr)
+			} else {
+				task = disabledTestTask(t, tc.taskName)
+			}
+			d.On("Task").Return(task)
+			d.On("TemplateIDs").Return(nil)
+			drivers.Add(tc.taskName, d)
+
+			controller := newTestController()
+			ctx := context.Background()
+
+			_, err := controller.checkApply(ctx, d, false, false)
+			data := controller.state.GetTaskEvents(tc.taskName)
+			events := data[tc.taskName]
+
+			if !tc.addToStore {
+				if tc.expectError {
+					assert.Error(t, err)
+					assert.Contains(t, err.Error(), tc.name)
+				}
+				assert.Equal(t, 0, len(events))
+				return
+			}
+
+			assert.Equal(t, 1, len(events))
+			event := events[0]
+			assert.Equal(t, tc.taskName, event.TaskName)
+			assert.False(t, event.StartTime.IsZero())
+			assert.False(t, event.EndTime.IsZero())
+
+			if tc.expectError {
+				assert.Error(t, err)
+				assert.Contains(t, err.Error(), tc.name)
+				assert.False(t, event.Success)
+				require.NotNil(t, event.EventError)
+				assert.Contains(t, event.EventError.Message, tc.name)
+			} else {
+				assert.NoError(t, err)
+				assert.True(t, event.Success)
+			}
+		})
+	}
+
+	t.Run("unrendered-scheduled-tasks", func(t *testing.T) {
+		// Test the behavior for once-mode and daemon-mode for the situation
+		// where a scheduled task's template did not render
+
+		controller := newTestController()
+
+		d := new(mocksD.Driver)
+		taskName := "scheduled_task"
+		d.On("Task").Return(scheduledTestTask(t, taskName))
+		d.On("TemplateIDs").Return(nil)
+		d.On("RenderTemplate", mock.Anything).Return(false, nil)
+		controller.drivers.Add(taskName, d)
+
+		// Once-mode - confirm no events are stored
+		ctx := context.Background()
+		_, err := controller.checkApply(ctx, d, false, true)
+		assert.NoError(t, err)
+		data := controller.state.GetTaskEvents(taskName)
+		events := data[taskName]
+		assert.Equal(t, 0, len(events))
+
+		// Daemon-mode - confirm an event is stored
+		_, err = controller.checkApply(ctx, d, false, false)
+		assert.NoError(t, err)
+		data = controller.state.GetTaskEvents(taskName)
+		events = data[taskName]
+		assert.Equal(t, 1, len(events))
+	})
+}
+
+func TestReadWrite_CheckApply_Store(t *testing.T) {
+	t.Run("mult-checkapply-store", func(t *testing.T) {
+		d := new(mocksD.Driver)
+		d.On("Task").Return(enabledTestTask(t, "task_a"))
+		d.On("TemplateIDs").Return(nil)
+		d.On("RenderTemplate", mock.Anything).Return(true, nil)
+		d.On("ApplyTask", mock.Anything).Return(nil)
+
+		disabledD := new(mocksD.Driver)
+		disabledD.On("Task").Return(disabledTestTask(t, "task_b"))
+		disabledD.On("TemplateIDs").Return(nil)
+
+		controller := newTestController()
+
+		controller.drivers.Add("task_a", d)
+		controller.drivers.Add("task_b", disabledD)
+		ctx := context.Background()
+
+		controller.checkApply(ctx, d, false, false)
+		controller.checkApply(ctx, disabledD, false, false)
+		controller.checkApply(ctx, d, false, false)
+		controller.checkApply(ctx, d, false, false)
+		controller.checkApply(ctx, d, false, false)
+		controller.checkApply(ctx, disabledD, false, false)
+
+		taskStatuses := controller.state.GetTaskEvents("")
+
+		assert.Equal(t, 4, len(taskStatuses["task_a"]))
+		assert.Equal(t, 0, len(taskStatuses["task_b"]))
+	})
+}
+
+func Test_once(t *testing.T) {
+	rw := &ReadWrite{}
+
+	testCases := []struct {
+		name     string
+		onceFxn  func(context.Context) error
+		numTasks int
+	}{
+		{
+			"consecutive one task",
+			rw.onceConsecutive,
+			1,
+		},
+		{
+			"consecutive multiple tasks",
+			rw.onceConsecutive,
+			10,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			errCh := make(chan error)
+			var errChRc <-chan error = errCh
+			go func() { errCh <- nil }()
+			w := new(mocksTmpl.Watcher)
+			w.On("WaitCh", mock.Anything).Return(errChRc)
+			w.On("Size").Return(tc.numTasks)
+
+			// Set up read-write controller with mocks
+			conf := multipleTaskConfig(tc.numTasks)
+			rw.baseController = &baseController{
+				state:   state.NewInMemoryStore(conf),
+				watcher: w,
+				drivers: driver.NewDrivers(),
+				newDriver: func(c *config.Config, task *driver.Task, w templates.Watcher) (driver.Driver, error) {
+					taskName := task.Name()
+					d := new(mocksD.Driver)
+					d.On("Task").Return(enabledTestTask(t, taskName)).Twice()
+					d.On("TemplateIDs").Return(nil)
+					d.On("RenderTemplate", mock.Anything).Return(false, nil).Once()
+					d.On("RenderTemplate", mock.Anything).Return(true, nil).Once()
+					d.On("InitTask", mock.Anything, mock.Anything).Return(nil).Once()
+					d.On("ApplyTask", mock.Anything).Return(nil).Once()
+					return d, nil
+				},
+				initConf: conf,
+				logger:   logging.NewNullLogger(),
+			}
+
+			ctx := context.Background()
+			err := rw.Init(ctx)
+			require.NoError(t, err)
+
+			// testing really starts here...
+			done := make(chan error)
+			// running in goroutine so I can timeout
+			go func() {
+				done <- tc.onceFxn(ctx)
+			}()
+			select {
+			case err := <-done:
+				assert.NoError(t, err, "unexpected error in Once()")
+			case <-time.After(time.Second):
+				t.Fatal("Once didn't return in expected time")
+			}
+
+			for _, d := range rw.drivers.Map() {
+				d.(*mocksD.Driver).AssertExpectations(t)
+			}
+		})
+	}
+}
+
+func TestReadWrite_deleteTask(t *testing.T) {
+	ctx := context.Background()
+	mockD := new(mocksD.Driver)
+
+	testCases := []struct {
+		name  string
+		setup func(*driver.Drivers)
+	}{
+		{
+			"success",
+			func(d *driver.Drivers) {
+				mockD.On("TemplateIDs").Return(nil)
+				d.Add("success", mockD)
+				mockD.On("Task").Return(enabledTestTask(t, "success"))
+				mockD.On("DestroyTask", ctx).Return()
+			},
+		},
+		{
+			"does_not_exist",
+			func(*driver.Drivers) {},
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			drivers := driver.NewDrivers()
+
+			tc.setup(drivers)
+			ctrl := ReadWrite{
+				baseController: &baseController{
+					logger: logging.NewNullLogger(),
+					state:  state.NewInMemoryStore(nil),
+				},
+			}
+			ctrl.baseController.drivers = drivers
+			ctrl.state.AddTaskEvent(event.Event{TaskName: "success"})
+
+			err := ctrl.deleteTask(ctx, tc.name)
+
+			assert.NoError(t, err)
+			_, exists := drivers.Get(tc.name)
+			assert.False(t, exists, "driver should no longer exist")
+
+			events := ctrl.state.GetTaskEvents(tc.name)
+			assert.Empty(t, events, "task events should no longer exist")
+
+			_, exists = ctrl.state.GetTask(tc.name)
+			assert.False(t, exists, "task should no longer exist in state")
+		})
+	}
+
+	t.Run("scheduled_task", func(t *testing.T) {
+		// Tests that deleting a scheduled task sends a stop notification
+
+		// Setup controller and drivers
+		taskName := "scheduled_task"
+		scheduledDriver := new(mocksD.Driver)
+		scheduledDriver.On("Task").Return(scheduledTestTask(t, taskName))
+		scheduledDriver.On("DestroyTask", ctx).Return()
+		scheduledDriver.On("TemplateIDs").Return(nil)
+		ctrl := newTestController()
+		ctrl.drivers.Add(taskName, scheduledDriver)
+		stopCh := make(chan struct{}, 1)
+		ctrl.scheduleStopChs[taskName] = stopCh
+
+		// Delete task
+		err := ctrl.deleteTask(ctx, taskName)
+		assert.NoError(t, err)
+
+		// Verify the stop channel received message
+		select {
+		case <-time.After(1 * time.Second):
+			t.Fatal("scheduled task was not notified to stop")
+		case <-stopCh:
+			break // expected case
+		}
+		_, ok := ctrl.scheduleStopChs[taskName]
+		assert.False(t, ok, "scheduled task stop channel still in map")
+	})
+
+	t.Run("active_task", func(t *testing.T) {
+		// Set up drivers with active task
+		drivers := driver.NewDrivers()
+		taskName := "active_task"
+		activeDriver := new(mocksD.Driver)
+		activeDriver.On("Task").Return(enabledTestTask(t, taskName))
+		activeDriver.On("DestroyTask", ctx).Return()
+		activeDriver.On("TemplateIDs").Return(nil)
+		drivers.Add(taskName, activeDriver)
+		drivers.SetActive(taskName)
+
+		// Set up controller with drivers and store
+		ctrl := ReadWrite{
+			baseController: &baseController{
+				logger: logging.NewNullLogger(),
+				state:  state.NewInMemoryStore(nil),
+			},
+		}
+		ctrl.baseController.drivers = drivers
+		ctrl.state.AddTaskEvent(event.Event{TaskName: taskName})
+
+		// Attempt to delete the active task
+		ch := make(chan error)
+		go func() {
+			err := ctrl.deleteTask(ctx, taskName)
+			ch <- err
+		}()
+
+		// Check that the task is not deleted while active
+		time.Sleep(500 * time.Millisecond)
+		_, exists := drivers.Get(taskName)
+		assert.True(t, exists, "task deleted when active")
+		events := ctrl.state.GetTaskEvents(taskName)
+		assert.NotEmpty(t, events, "task events should still exist")
+
+		// Set task to inactive, wait for deletion to happen
+		drivers.SetInactive(taskName)
+		select {
+		case err := <-ch:
+			assert.NoError(t, err)
+		case <-time.After(1 * time.Second):
+			t.Fatal("task was not deleted after it became inactive")
+		}
+
+		// Check that task removed from drivers and store
+		_, exists = drivers.Get(taskName)
+		assert.False(t, exists, "driver should no longer exist")
+
+		events = ctrl.state.GetTaskEvents(taskName)
+		assert.Empty(t, events, "task events should no longer exist")
+
+		_, exists = ctrl.state.GetTask(taskName)
+		assert.False(t, exists, "task should no longer exist in state")
+	})
+
+}
+func TestReadWrite_waitForTaskInactive(t *testing.T) {
+	ctx := context.Background()
+	t.Run("active_task", func(t *testing.T) {
+		// Set up drivers with active task
+		ctrl := newTestController()
+		taskName := "inactive_task"
+		mockD := new(mocksD.Driver)
+		mockD.On("TemplateIDs").Return(nil)
+		ctrl.drivers.Add(taskName, mockD)
+		ctrl.drivers.SetActive(taskName)
+
+		// Wait for task to become inactive
+		ch := make(chan error)
+		go func() {
+			err := ctrl.waitForTaskInactive(ctx, taskName)
+			ch <- err
+		}()
+
+		// Check that the wait does not return early
+		select {
+		case <-ch:
+			t.Fatal("wait completed when task was still active")
+		case <-time.After(250 * time.Millisecond):
+			break
+		}
+
+		// Set task to inactive, wait should complete
+		ctrl.drivers.SetInactive(taskName)
+		select {
+		case err := <-ch:
+			assert.NoError(t, err)
+		case <-time.After(1 * time.Second):
+			t.Fatal("wait should have completed because task is inactive")
+		}
+	})
+
+	t.Run("inactive_task", func(t *testing.T) {
+		// Set up drivers with inactive task
+		ctrl := newTestController()
+		taskName := "inactive_task"
+		mockD := new(mocksD.Driver)
+		mockD.On("TemplateIDs").Return(nil)
+		ctrl.drivers.Add(taskName, mockD)
+		ctrl.drivers.SetInactive(taskName)
+
+		// Wait for task to be inactive, should return immediately
+		err := ctrl.waitForTaskInactive(ctx, taskName)
+		require.NoError(t, err)
 	})
 }
